Move shared content I/O helpers from content/io.ts to utils/io.ts
content/ should read as a list of content types, not a grab bag of
infrastructure. The bufferToStream + read/stream Zod schemas are
helpers consumed by the content types, so they belong in utils/
alongside the other cross-content helpers (vcard, audio, stream).
